fix dockerfile

+++ entrypoint.sh
@@ -1,13 +1,14 @@
-#!/bin/sh
+#!/bin/bash
 
 : ${LABO_USER:=labo}
 
-adduser -S $LABO_USER && \
-  addgroup $LABO_USER && \
-  usermod -g $LABO_USER -G $LABO_USER,sudo $LABO_USER
+adduser -D $LABO_USER && \
+  usermod -g $LABO_USER -G $LABO_USER,sudo -s /bin/zsh $LABO_USER
 
 if [ -n "$LABO_TIMEZONE" ]; then
-  ln -sf /usr/share/zoneinfo/$LABO_TIMEZONE /etc/localtime
+  if [ -f "/usr/share/zoneinfo/$LABO_TIMEZONE" ]; then
+    ln -sf /usr/share/zoneinfo/$LABO_TIMEZONE /etc/localtime
+  fi
 fi
 
-exec sudo -u $LABO_USER "$@"
+exec sudo -H -u $LABO_USER "$@"
